defect_detection_project/requirements.txt
# 瑕疵检测项目依赖

# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# 数据增强
albumentations>=1.3.0
opencv-python>=4.7.0

# 训练工具
tqdm>=4.65.0
scikit-learn>=1.2.0

# 可视化
matplotlib>=3.7.0

# 可选：日志和实验跟踪
# tensorboard>=2.13.0
# wandb>=0.15.0

# 可选：分割损失函数
# segmentation-models-pytorch>=0.3.3

# 注意：DINOv3 会通过 torch.hub 自动下载，无需手动安装
